# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "3&7 Training Platform API",
    "description": "API for the 3&7 Training and Recovery Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/reports/training": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Training Report",
        "description": "Generate a training report for the specified date range.\n\nThe report includes:\n- Summary of training sessions\n- List of sessions with details\n- Attendance stats for each session\n- Feedback summary for each session\n\nReturns:\n- JSON data by default\n- PDF or PowerPoint if format is specified",
        "operationId": "get_training_report_reports_training_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Start date in format YYYY-MM-DD",
              "title": "Start Date"
            },
            "description": "Start date in format YYYY-MM-DD"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "End date in format YYYY-MM-DD",
              "title": "End Date"
            },
            "description": "End date in format YYYY-MM-DD"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReportFormat"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Format of the report (json, pdf, ppt)",
              "default": "json",
              "title": "Format"
            },
            "description": "Format of the report (json, pdf, ppt)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingReportResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/attendance": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Attendance Report",
        "description": "Generate an attendance report for the specified date range.\n\nThe report includes:\n- Overall attendance statistics\n- For a team: attendance rate for each athlete\n- For an individual: sessions attended/missed with details\n\nReturns:\n- JSON data by default\n- PDF or PowerPoint if format is specified",
        "operationId": "get_attendance_report_reports_attendance_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Start date in format YYYY-MM-DD",
              "title": "Start Date"
            },
            "description": "Start date in format YYYY-MM-DD"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "End date in format YYYY-MM-DD",
              "title": "End Date"
            },
            "description": "End date in format YYYY-MM-DD"
          },
          {
            "name": "athlete_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Get attendance for a specific athlete",
              "title": "Athlete Id"
            },
            "description": "Get attendance for a specific athlete"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReportFormat"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Format of the report (json, pdf, ppt)",
              "default": "json",
              "title": "Format"
            },
            "description": "Format of the report (json, pdf, ppt)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceReportResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/feedback": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Feedback Report",
        "description": "Generate a feedback report for the specified date range.\n\nThe report includes:\n- Overall feedback statistics\n- For a date range: feedback summary for each session\n- For a specific session: detailed feedback from each athlete\n\nReturns:\n- JSON data by default\n- PDF or PowerPoint if format is specified",
        "operationId": "get_feedback_report_reports_feedback_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Start date in format YYYY-MM-DD",
              "title": "Start Date"
            },
            "description": "Start date in format YYYY-MM-DD"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "End date in format YYYY-MM-DD",
              "title": "End Date"
            },
            "description": "End date in format YYYY-MM-DD"
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Get feedback for a specific session",
              "title": "Session Id"
            },
            "description": "Get feedback for a specific session"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReportFormat"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Format of the report (json, pdf, ppt)",
              "default": "json",
              "title": "Format"
            },
            "description": "Format of the report (json, pdf, ppt)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackReportResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/export": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Export Report",
        "description": "Export a custom report as PDF or PowerPoint.\n\nAccepts:\n- Custom report data\n- Title for the report\n- Format (PDF or PPT)\n\nReturns:\n- PDF or PowerPoint file",
        "operationId": "export_report_reports_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/reports/monthly": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Monthly Report",
        "description": "Generate a monthly report including both training sessions and independent training\n\nThis combines data from both training sessions and independent training sessions\nto provide a comprehensive monthly view for coaches and athletes.",
        "operationId": "get_monthly_report_reports_monthly_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Start date in format YYYY-MM-DD",
              "title": "Start Date"
            },
            "description": "Start date in format YYYY-MM-DD"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "End date in format YYYY-MM-DD",
              "title": "End Date"
            },
            "description": "End date in format YYYY-MM-DD"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReportFormat"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Format of the report (json, pdf, ppt)",
              "default": "json",
              "title": "Format"
            },
            "description": "Format of the report (json, pdf, ppt)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyReportResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/overview": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Overview Report",
        "description": "Get overview report with statistics",
        "operationId": "get_overview_report_reports_overview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttendanceReportResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "date_range": {
            "type": "string",
            "title": "Date Range"
          },
          "summary": {
            "type": "object",
            "title": "Summary"
          },
          "data": {
            "items": {},
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "title",
          "generated_at",
          "date_range",
          "summary",
          "data"
        ],
        "title": "AttendanceReportResponse",
        "description": "Response model for attendance report"
      },
...
      "FeedbackReportResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "date_range": {
            "type": "string",
            "title": "Date Range"
          },
          "summary": {
            "type": "object",
            "title": "Summary"
          },
          "data": {
            "items": {},
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "title",
          "generated_at",
          "date_range",
          "summary",
          "data"
        ],
        "title": "FeedbackReportResponse",
        "description": "Response model for feedback report"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportFormat": {
        "type": "string",
        "enum": [
          "json",
          "pdf",
          "ppt"
        ],
        "title": "ReportFormat",
        "description": "Enum for report formats"
      },
...
      "TrainingReportResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "date_range": {
            "type": "string",
            "title": "Date Range"
          },
          "summary": {
            "type": "object",
            "title": "Summary"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/TrainingSessionSummary"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "title",
          "generated_at",
          "date_range",
          "summary",
          "data"
        ],
        "title": "TrainingReportResponse",
        "description": "Response model for training report"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth/token"
          }
        }
      }
    }
  }
}